Docs: set up routing, added /docs url
