+++ docs/src/components/Jumbotron/styles.module.scss
@@ -1,4 +1,3 @@
-
 .container {
   display: flex;
   flex-flow: column;
